requirements.txt
# Core Requirements
Flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
Pillow==10.0.1
plotly==5.18.0
pandas==2.0.3
numpy==1.24.4

# Database
pymongo==4.6.1
motor==3.3.2

# AI and Machine Learning
scikit-learn==1.3.2
scipy==1.10.1

# Audio Processing
SpeechRecognition==3.10.0
pydub==0.25.1
python-Levenshtein==0.21.1

# Visualization
matplotlib==3.7.3

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Utility
python-multipart==0.0.6
tqdm==4.66.1
pytest==7.4.3

# Deployment
uvicorn==0.23.2
gunicorn==21.2.0

# Web Framework
itsdangerous==2.1.2
click==8.1.7